all_unique {dendextend}	R Documentation

Check if all the elements in a vector are unique

Description

Checks if all the elements in a vector are unique

Usage

all_unique(x, ...)

Arguments

x
a vector

...
ignored.

Value

logical (are all the elements in the vector unique)

Source

http://r.789695.n4.nabble.com/Is-there-a-function-to-test-if-all-the-elements-in-a-vector-are-unique-td931833.html

See Also

unique

Examples

all_unique(c(1:5, 1,1))
all_unique(c(1,1,2))
all_unique(c(1,1,2, 3,3,3,3))
all_unique(c(1,3,2))
all_unique(c(1:10))

[Package dendextend version 1.9.0 Index]