as.dendlist {dendextend}	R Documentation

Try to coerce something into a dendlist

Description

It removes stuff that are not dendgrogram/dendlist and turns what is left into a dendlist

Usage

as.dendlist(x, ...)

Arguments

x
a list with several dendrogram/hclust/phylo or dendlist objects and other junk that should be omitted.

...
NOT USED

Value

A list of class dendlist where each item is a dendrogram

Examples

## Not run: 

dend <- iris[,-5] %>% dist %>% hclust %>% as.dendrogram
dend2 <- iris[,-5] %>% dist %>% hclust(method = "single") %>% as.dendrogram

x = list(dend, 1, dend2)
as.dendlist(x)

## End(Not run)

[Package dendextend version 1.9.0 Index]