bakers_gamma_for_2_k_matrix {dendextend}	R Documentation

Bakers Gamma for two k matrices

Description

Bakers Gamma for two k matrices

Usage

bakers_gamma_for_2_k_matrix(k_matrix_dend1, k_matrix_dend2, to_plot = FALSE)

Arguments

k_matrix_dend1
a matrix of k cluster groupings from a dendrogram

k_matrix_dend2
a (second) matrix of k cluster groupings from a dendrogram

to_plot
logical (FALSE). Should a scaterplot be plotted, showing the correlation between the lowest shared branch between two items in the two compared trees.

...
not used

Value

Baker's Gamma coefficient.

See Also

cor_bakers_gamma

[Package dendextend version 1.9.0 Index]