Interactively rotate a tree object

Lets te user click a plot of dendrogram and rotates the tree based on the location of the click.

Code for mouse selection of (sub-)cluster to be rotated

click_rotate(x, ...)

# S3 method for default
click_rotate(x, ...)

# S3 method for dendrogram
click_rotate(
  x,
  plot = TRUE,
  plot_after = plot,
  horiz = FALSE,
  continue = FALSE,
  ...
)

Arguments

x

a tree object (either a dendrogram or hclust)

...

parameters passed to the plot

plot

(logical) should the dendrogram first be plotted.

plot_after

(logical) should the dendrogram be plotted after the rotation?

horiz

logical. Should the plot be normal or horizontal?

continue

logical. If TRUE, allows the user to keep clicking the plot until a click is made on the labels.

Value

A rotated tree object

See also

Author

Andrej-Nikolai Spiess, Tal Galili

Examples

# create the dend:
dend <- USArrests %>%
  dist() %>%
  hclust("ave") %>%
  as.dendrogram() %>%
  color_labels()
if (FALSE) {
# play with the rotation once
dend <- click_rotate(dend)
dend <- click_rotate(dend, horiz = TRUE)
# keep playing with the rotation:
while (TRUE) dend <- click_rotate(dend)
# the same as
dend <- click_rotate(dend, continue = TRUE)
}