Cut a dendrogram - and run a function on the output

Cuts the dend at height h and returns a list with the FUN function implemented on all the sub trees created by cut at height h. This is used for creating a cutree.dendrogram function, by using the labels function as FUN.

This is the Rcpp version of the function, offering a 10-60 times improvement in speed (depending on the tree size it is used on).

cut_lower_fun(dend, h, FUN = labels, warn = dendextend_options("warn"), ...)

Arguments

dend
a dendrogram object.

h
a scalar of height to cut the dend by.

FUN
a function to run. (default is "labels")

warn
logical (default from dendextend_options("warn") is FALSE). Set if warning are to be issued, it is safer to keep this at TRUE, but for keeping the noise down, the default is FALSE. Should the user be warned if reverting to default?

...
passed to FUN.

Value

A list with the output of running FUN on each of the sub dends derived from cutting "dend"

See also

Examples

dend <- as.dendrogram(hclust(dist(iris[1:4, -5])))
# this is really cool!
cut_lower_fun(dend, .4, labels)

#> [[1]]
#> [1] "1"
#> 
#> [[2]]
#> [1] "2" "3" "4"
#> 

lapply(cut(dend, h = .4)$lower, labels)

#> [[1]]
#> [1] "1"
#> 
#> [[2]]
#> [1] "2" "3" "4"
#> 

cut_lower_fun(dend, .4, order.dendrogram)

#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 2 3 4
#> 

if (FALSE) {
# library(dendextend)
library(dendextendRcpp)
dend_big <- as.dendrogram(hclust(dist(iris[1:150, -5])))
library(microbenchmark)
microbenchmark(old_cut_lower_fun(dend_big, .1),
  dendextendRcpp::dendextendRcpp_cut_lower_fun(dend_big, .1),
  times = 100
)
# about 7-15 times faster. It is faster the larger the tree is, and the lower h is.
}