d3dendro_options {dendextend}	R Documentation

Get or set current defaults for d3dendrogram

Description

Get or set current defaults for d3dendrogram

Usage

d3dendro_options(..., copy = FALSE)

Arguments

...
<option>=<value> pairs (see below). If no option/value pair is given, a list with current settings is returned. If the single input "reset" is given, options are restored to factory settings.

copy
only when options are set: set global options or return a modified copy of the global option environment invisibly?

Options

Option values are exported as-is to the generated javascript/webpage.

node_fill (char,"#fff") 	 Node color (fill)

node_stroke (char;"steelblue") 	 Node color (circumference)

node_stroke_width (char;"1.5px") 	 Width of node circumference

node_radius (num;4.5) 	 Size of node (pixels)

node_font (char;"14px sans-serif") 	 Font used to print labels. Size and type spec

link_fill (char;"none") 	 link fill color (you probably do not want to set this)

link_stroke (char;"#ccc") 	 link line color (that's the one you want)

link_stroke_width (char;"1.5px") 	 link line width (pixels)

axis_stroke (char;"black") 	 Color of axis line

axis_width (char; "2px") 	 Width of axis line

axis_ntick (num; 5) 	 Approximate nr of tick marks

axis_ticklength_in (num;5) 	 Length of tickmarks into the graph (pixels)

axis_ticklength_out (num;5) 	 Length of tickmarks out of the graph (pixels)

axis_tickmark_offset (num;8) 	 Distance between axis and tick labels (pixels)

[Package dendextend version 1.9.0 Index]