dendextend-package {dendextend}	R Documentation

Functions for extending dendrogram objects

Description

Offers a set of functions for extending 'dendrogram' objects in R, letting you visualize and compare trees of 'hierarchical clusterings'. You can (1) Adjust a tree's graphical parameters - the color, size, type, etc of its branches, nodes and labels. (2) Visually and statistically compare different 'dendrograms' to one another.

Author(s)

Maintainer: Tal Galili tal.galili@gmail.com (https://www.r-statistics.com) [copyright holder]

Authors:

Gregory Jefferis jefferis@gmail.com (imported code from his dendroextras package) [contributor]

Other contributors:

Yoav Benjamini ybenja@tau.ac.il [thesis advisor]

Gavin Simpson [contributor]

Marco Gallotta (a.k.a: marcog) [contributor]

Johan Renaudie (https://github.com/plannapus) [contributor]

The R Core Team (Thanks for the Infastructure, and code in the examples) [contributor]

Kurt Hornik [contributor]

Uwe Ligges [contributor]

Andrej-Nikolai Spiess [contributor]

Steve Horvath SHorvath@mednet.ucla.edu [contributor]

Peter Langfelder Peter.Langfelder@gmail.com [contributor]

skullkey [contributor]

Mark Van Der Loo mark.vanderloo@gmail.com (https://github.com/markvanderloo d3dendrogram) [contributor]

Andrie de Vries apdevries@gmail.com (ggdendro author) [contributor]

Zuguang Gu z.gu@dkfz-heidelberg.de (circlize author) [contributor]

Cath (https://github.com/CathG) [contributor]

John Ma (https://github.com/JohnMCMa) [contributor]

Krzysiek G (https://github.com/storaged) [contributor]

Manuela Hummel m.hummel@dkfz.de (https://github.com/hummelma) [contributor]

Chase Clark (https://github.com/chasemc) [contributor]

Lucas Graybuck (https://github.com/hypercompetent) [contributor]

jdetribol (https://github.com/jdetribol) [contributor]

Ben Ho ben.ho@sickkids.ca (https://github.com/SplitInf) [contributor]

Samuel Perreault samuel.perreault.3@ulaval.ca (https://github.com/samperochkin) [contributor]

Christian Hennig c.hennig@ucl.ac.uk (http://www.homepages.ucl.ac.uk/~ucakche/) [contributor]

David Bradley (https://github.com/DBradley27) [contributor]

Houyun Huang houyunhuang@163.com (https://github.com/houyunhuang) [contributor]

Patrick Schupp pschupp@sonic.net (https://github.com/pschupp) [contributor]

Alec Buetow alecbuetow@gmail.com (https://github.com/alecbuetow) [contributor]

See Also

dendrogram, hclust in stats package.

[Package dendextend version 1.19.1 Index]