dendextend_labels.dendrogram {dendextend}	R Documentation

Labels of the Leaves in a Dendrogram

Description

This is the same function as labels.dendrogram. It is used inside dendextend_options, and thus allows the function to be updated when loading the dendextendRcpp package.

Usage

dendextend_labels.dendrogram(object, ...)

Arguments

object
a dendrogram.

...
not used.

Value

A vector with length equal to the number of leaves in the dendrogram is returned. From r <- order.dendrogram(), each element is the index into the original data (from which the dendrogram was computed).

See Also

labels.dendrogram

Examples

# stats:::labels.dendrogram

[Package dendextend version 1.7.0 Index]