dendextend_options {dendextend}	R Documentation

Access to dendextend_options

Description

This is a function inside its own environment. This enables a bunch of functions to be manipulated outside the package, even when they are called from function within the dendextend package.

TODO: describe options.

A new "warn" dendextend_options parameter. logical (FALSE). Should warning be issued?

Usage

dendextend_options(option, value)

Arguments

option
a character scalar of the value of the options we would like to access or update.

value
any value that we would like to update into the "option" element in dendextend_options

Value

a list with functions

Author(s)

Kurt Hornik

Examples

dendextend_options("a")
dendextend_options("a", 1)
dendextend_options("a")
dendextend_options("a", NULL)
dendextend_options("a")
dendextend_options()

[Package dendextend version 1.9.0 Index]