Duplicate a leaf X times

Duplicates a leaf in a tree. Useful for non-parametric bootstraping trees since it emulates what would have happened if the tree was constructed based on a row-sample with replacments from the original data matrix.

duplicate_leaf(
  dend,
  leaf_label,
  times,
  fix_members = TRUE,
  fix_order = TRUE,
  fix_midpoint = TRUE,
  ...
)

Arguments

dend
a dendrogram object

leaf_label
the label of the laef to replicate.

times
the number of times we will have this leaf after replication

fix_members
logical (TRUE). Fix the number of members in attr using fix_members_attr.dendrogram

fix_order
logical (TRUE). Fix the leaves order

fix_midpoint
logical (TRUE). Fix the midpoint value. If TRUE, it overrides "fix_members" and turns it into TRUE (since it must have a correct number of members in order to work). values using rank_order.dendrogram

...
not used

Value

A dendrogram, after duplicating one of its leaves.

Examples

if (FALSE) {
# define dendrogram object to play with:
dend <- USArrests[1:3, ] %>%
  dist() %>%
  hclust(method = "ave") %>%
  as.dendrogram()
plot(dend)
duplicate_leaf(dend, "Alaska", 3)
duplicate_leaf(dend, "Arizona", 2, fix_members = FALSE, fix_order = FALSE)
plot(duplicate_leaf(dend, "Alaska", 2))
plot(duplicate_leaf(dend, "Alaska", 4))
plot(duplicate_leaf(dend, "Arizona", 2))
plot(duplicate_leaf(dend, "Arizona", 4))
}