flip_leaves {dendextend}	R Documentation

Flip leaves

Description

Rotate a branch in a tree so that the locations of two bundles of leaves are flipped.

Usage

flip_leaves(dend, leaves1, leaves2, ...)

Arguments

dend
a dendrogram object

leaves1
a vector of leaves order value to flip.

leaves2
a (second) vector of leaves order value to flip.

...
not used

Details

This function is based on a bunch of string manipulation functions. There may be a smarter/better way for doing it...

Value

A dendrogram object with flipped leaves.

See Also

tanglegram, match_order_by_labels, entanglement.

Examples

## Not run: 
dend1 <- USArrests[1:5,] %>% dist %>% hclust %>% as.dendrogram
dend2 <- flip_leaves(dend1, c(3,5), c(1,2))
tanglegram(dend1,dend2)
entanglement(dend1,dend2, L = 2) # 0.4

## End(Not run)

[Package dendextend version 1.9.0 Index]