get_root_branches_attr {dendextend}	R Documentation

get attributes from the dendrogram's root(!) branches

Description

get attributes from the dendrogram's root(!) branches

Usage

get_root_branches_attr(dend, the_attr, warn = dendextend_options("warn"), ...)

Arguments

dend
dendrogram object

the_attr
the attribute to get from the branches (for example "height")

warn
logical (default from dendextend_options("warn") is FALSE). Set if warning are to be issued, it is safer to keep this at TRUE, but for keeping the noise down, the default is FALSE. Should a warning be printed when the function is used on an object which is NOT a dendrogram.

...
passed on to attr

Value

The attributes of the branches (often two) of the dendrogram's root

See Also

attr

Examples

hc <- hclust(dist(USArrests[2:9,]), "com")
dend <- as.dendrogram(hc)

get_root_branches_attr(dend, "height") # 0.00000 71.96247
# plot(dend)
str(dend, 2)

[Package dendextend version 1.9.0 Index]