lowest_common_branch {dendextend}	R Documentation

Find lowest common branch were the two items are shared

Description

Given two vectors, for two items, of cluster belonging - the function finds the lowest branch (e.g: largest number of k clusters) for which the two items are in the same cluster for the two trees.

Usage

lowest_common_branch(item1, item2, ...)

Arguments

item1
a named numeric vector (of cluster group with names of k level)

item2
a named numeric vector (of cluster group with names of k level)

...
not used

Value

The first location (from left) where the two vectors have the same A dendrogram, after adjusting the members attr in all of its nodes.

See Also

cor_bakers_gamma

Examples

item1 <- structure(c(1L, 1L, 1L, 1L), .Names = c("1", "2", "3", "4"))
item2 <- structure(c(1L, 1L, 2L, 2L), .Names = c("1", "2", "3", "4"))
lowest_common_branch(item1, item2)

[Package dendextend version 1.9.0 Index]