match_order_by_labels {dendextend}	R Documentation

Adjust the order of one dendrogram based on another (using labels)

Description

Takes one dendrogram and adjusts its order leaves valeus based on the order of another dendrogram. The values are matached based on the labels of the two dendrograms.

This allows for faster entanglement running time, since we can be sure that the leaves order is just as using their labels.

Usage

match_order_by_labels(dend_change, dend_template,
  check_that_labels_match = TRUE)

Arguments

dend_change
tree object (dendrogram)

dend_template
tree object (dendrogram)

check_that_labels_match
logical (TRUE). If to check that the labels in the two dendrogram match. (if they do not, the function aborts)

Value

Returns dend_change after adjusting its order values to be like dend_template.

See Also

entanglement , tanglegram

Examples

## Not run: 

dend <- USArrests[1:4,] %>% dist %>% hclust %>% as.dendrogram
order.dendrogram(dend) #  c(4L, 3L, 1L, 2L)

dend_changed <- dend
order.dendrogram(dend_changed) <- 1:4
order.dendrogram(dend_changed) # c(1:4)

# now let's fix the order of the new object to be as it was:
dend_changed <- match_order_by_labels(dend_changed, dend)
# these two are now the same:
order.dendrogram(dend_changed)
order.dendrogram(dend))

## End(Not run)

[Package dendextend version 1.9.0 Index]