prune {dendextend}	R Documentation

Prunes a tree (using leaves' labels)

Description

Trimms a tree (dendrogram, hclust) from a set of leaves based on their labels.

Usage

prune(dend, ...)

## Default S3 method:
prune(dend, ...)

## S3 method for class 'dendrogram'
prune(dend, leaves, reindex_dend = TRUE, ...)

## S3 method for class 'hclust'
prune(dend, leaves, ...)

## S3 method for class 'phylo'
prune(dend, ...)

## S3 method for class 'rpart'
prune(dend, ...)

Arguments

dend
tree object (dendrogram/hclust/phylo)

...
passed on

leaves
a character vector of the label(S) of the tip(s) (leaves) we wish to prune off the tree.

reindex_dend
logical (default is TRUE). If TRUE, the leaves of the new dendrograms include the rank of the old order.dendrogram. This insures that their values are just like the number of leaves. When FALSE, the values in the leaves is that of the original dendrogram. Thie is useful if prunning a dendrogram but then wanting to use order.dendrogram with the original values. When using prune.hclust, then reindex_dend is used by default since otherwise the as.hclust function would return an error.

Details

I was not sure if to call this function drop.tip (from ape), snip/prune (from rpart) or just remove.leaves. I ended up deciding on prune.

Value

A pruned tree

See Also

prune_leaf, drop.tip ape

Examples

hc <- hclust(dist(USArrests[1:5,]), "ave")
dend <- as.dendrogram(hc)

par(mfrow = c(1,2))
plot(dend, main = "original tree")
plot(prune(dend , c("Alaska", "California")), main = "tree without Alaska and California")

# this works because prune uses reindex_dend = TRUE by default
as.hclust(prune(dend , c("Alaska", "California")))
prune(hc , c("Alaska", "California"))

[Package dendextend version 1.9.0 Index]