prune_common_subtrees.dendlist {dendextend}	R Documentation

Prune trees to their common subtrees

Description

Prune trees to their common subtrees

Usage

prune_common_subtrees.dendlist(dend, ...)

Arguments

dend
a dendlist of length two

...
ignored

Value

A dendlist after prunning the labels to only include those that are part of common subtrees in both dendrograms.

See Also

common_subtrees_clusters

Examples

# NULL

[Package dendextend version 1.9.0 Index]