pvclust_edges {dendextend}	R Documentation

Get Pvclust Edges Information

Description

Get pvclust edges information such as au and bp and return dataframe with proper sample labels. This function is useful when there are a lot of samples involved.

Usage

pvclust_edges(pvclust_obj)

Arguments

pvclust_obj
pvclust object

Value

data.frame with leaves on column 1 and 2, followed by the rest of the information from edge

References

hclust object descriptions https://stat.ethz.ch/R-manual/R-patched/library/stats/html/hclust.html

Examples

## Not run: 

library(pvclust)
data(lung) # 916 genes for 73 subjects
set.seed(13134)
result <- pvclust(lung[, 1:20], method.dist="cor", method.hclust="average", nboot=100)
pvclust_edges(result)

## End(Not run)

[Package dendextend version 1.10.0 Index]