raise.dendrogram {dendextend}	R Documentation

Raise the height of a dendrogram tree

Description

Raise the height of nodes in a dendrogram tree.

Usage

raise.dendrogram(dend, heiget_to_add, ...)

Arguments

dend
dendrogram object

heiget_to_add
how much height to add to all the branches (not leaves) in the dendrogram

...
passed on (not used)

Value

A raised dendrogram

Examples

hc <- hclust(dist(USArrests[2:9,]), "com")
dend <- as.dendrogram(hc)

par(mfrow = c(1,2))
plot(dend, main = "original tree")
plot(raise.dendrogram(dend , 100), main = "Raised tree")

[Package dendextend version 1.9.0 Index]