Rotates dend based on DendSer

Rotates a dendrogram based on its seriation

The function tries to turn the dend into hclust using DendSer.dendrogram (based on DendSer)

Also, if a distance matrix is missing, it will try to use the cophenetic distance.

rotate_DendSer(dend, ser_weight, ...)

Arguments

dend

An object of class dendrogram

ser_weight

Used by cost function to evaluate ordering. For cost=costLS, this is a vector of object weights. Otherwise is a dist or symmetric matrix. passed to DendSer.dendrogram and from there to DendSer.

If it is missing, the cophenetic distance is used instead.

...

parameters passed to DendSer

Value

Numeric vector giving an optimal dendrogram order

See also

Examples

if (FALSE) { # \dontrun{
library(DendSer) # already used from within the function

dend <- USArrests[1:4, ] %>%
  dist() %>%
  hclust("ave") %>%
  as.dendrogram()
DendSer.dendrogram(dend)

tanglegram(dend, rotate_DendSer(dend))
} # }