seriate_dendrogram {dendextend}	R Documentation

Rotates a dendrogram based on a seriation of a distance matrix

Description

Rotates a dendrogram so it confirms to an order of a provided distance object. The seriation algorithm is based on seriate, which tries to find a linear order for objects using data in form of a dissimilarity matrix (one mode data).

This is useful for heatmap visualization.

Usage

seriate_dendrogram(dend, x, method = c("OLO", "GW"), ...)

Arguments

dend
An object of class dendrogram or hclust

x
a dist object.

method
a character vector of either "OLO" or "GW": "OLO" - Optimal leaf ordering, optimzes the Hamiltonian path length that is restricted by the dendrogram structure - works in O(n^4) "GW" - Gruvaeus and Wainer heuristic to optimze the Hamiltonian path length that is restricted by the dendrogram structure

...
parameters passed to seriate

Value

A dendrogram that is rotated based on the optimal ordering of the distance matrix

See Also

rotate, seriate

Examples

## Not run: 
# library(dendextend)
d <- dist(USArrests)
hc <- hclust(d, "ave")
dend <- as.dendrogram(hc)

heatmap(as.matrix(USArrests))

dend2 <- seriate_dendrogram(dend, d)
heatmap(as.matrix(USArrests), Rowv = dend)

## End(Not run)

[Package dendextend version 1.9.0 Index]