sort_dist_mat {dendextend}	R Documentation

Sorts a distance matrix by rows and columns names

Description

Sorts a distance matrix by the names of the rows and columns.

Usage

sort_dist_mat(dist_mat, by_rows = TRUE, by_cols = TRUE, ...)

Arguments

dist_mat
a distance matrix.

by_rows
logical (TRUE). Sort the distance matrix by rows?

by_cols
logical (TRUE). Sort the distance matrix by columns?

...
Ignored.

Value

A distance matrix (after sorting)

See Also

dist, cor_cophenetic

[Package dendextend version 1.9.0 Index]