theme_dendro {dendextend}	R Documentation

Creates completely blank theme in ggplot

Description

Sets most of the ggplot options to blank, by returning blank theme elements for the panel grid, panel background, axis title, axis text, axis line and axis ticks.

Usage

theme_dendro()

Author(s)

Andrie de Vries

Source

This function is from Andrie de Vries's ggdendro package.

The motivation for this fork is the need to add more graphical parameters to the plotted tree. This required a strong mixter of functions from ggdendro and dendextend (to the point that it seemed better to just fork the code into its current form)

See Also

ggdend

[Package dendextend version 1.9.0 Index]