unbranch {dendextend}	R Documentation

unbranch trees

Description

unbranch trees and merges the subtree to the parent node.

Usage

unbranch(dend, ...)

## Default S3 method:
unbranch(dend, ...)

## S3 method for class 'dendrogram'
unbranch(dend, branch_becoming_root = 1, new_root_height,
  ...)

## S3 method for class 'hclust'
unbranch(dend, branch_becoming_root = 1, new_root_height,
  ...)

## S3 method for class 'phylo'
unbranch(dend, ...)

Arguments

dend
a dendrogram (or hclust) object

...
passed on

branch_becoming_root
a numeric choosing the branch of the root which will become the new root (from left to right)

new_root_height
the new height of the branch which will become the new root. If the parameter is not given - the height of the original root is used.

Value

An unbranched dendrogram

See Also

unroot ape

Examples

hc <- hclust(dist(USArrests[2:9,]), "com")
dend <- as.dendrogram(hc)

par(mfrow = c(1,3))
plot(dend, main = "original tree")
plot(unbranch(dend , 1), main = "unbranched tree (left branch)")
plot(unbranch(dend , 2), main = "tree without  (right branch)")

[Package dendextend version 1.9.0 Index]