untangle_DendSer {dendextend}	R Documentation

Tries to run DendSer on a dendrogram

Description

The function tries to turn the dend into hclust. It then uses the cophenetic distance matrix for optimizing the tree's rotation.

This is a good (and fast) starting point for linkuntangle_step_rotate_2side

Usage

untangle_DendSer(dend, ...)

Arguments

dend
An object of class dendlist

...
NOT USED

Value

A dendlist object with ordered dends

See Also

DendSer, DendSer.dendrogram , untangle_DendSer, rotate_DendSer

Examples

## Not run: 
set.seed(232)
ss <- sample(1:150, 20)
dend1 <- iris[ss,-5] %>% dist %>% hclust("com") %>% as.dendrogram
dend2 <- iris[ss,-5] %>% dist %>% hclust("sin") %>% as.dendrogram
dend12 <- dendlist(dend1, dend2)

# bad solutions
dend12 %>% tanglegram
dend12 %>% untangle("step2") %>% tanglegram
dend12 %>% untangle_DendSer %>% tanglegram
# but the combination is quite awsome:
dend12 %>% untangle_DendSer %>% untangle("step2") %>% tanglegram

## End(Not run)

[Package dendextend version 1.9.0 Index]