Change col/lwd/lty of branches matching labels condition

The user supplies a dend, labels, and type of condition (all/any), and TF_values And the function returns a dendrogram with branches col/lwd/lty accordingly

branches_attr_by_labels(dend, labels, TF_values = c(2, Inf),
  attr = c("col", "lwd", "lty"), type = c("all", "any"), ...)

Arguments

dend
a dendrogram dend

labels
a character vector of labels from the tree

TF_values
a two dimensional vector with the TF_values to use in case a branch fulfills the condition (TRUE) and in the case that it does not (FALSE). Defaults are 2/Inf for col, lwd and lty. (so it will insert the first value, and will not change all the FALSE cases)

attr
a character with one of the following values: col/lwd/lty

type
a character vector of either "all" or "any", indicating which of the branches should be painted: ones that all of their labels belong to the supplied labels, or also ones that even some of their labels are included in the labels vector.

...
ignored.

Value

A dendrogram with modified branches (col/lwd/lty).

See also

Examples

# NOT RUN {
library(dendextend)

set.seed(23235)
ss <- sample(1:150, 10 )

# Getting the dend dend
dend <- iris[ss,-5] %>% dist %>% hclust %>% as.dendrogram
dend %>% plot

dend %>%
   branches_attr_by_labels(c("123", "126", "23", "29")) %>%
   plot
dend %>%
   branches_attr_by_labels(c("123", "126", "23", "29"), "all") %>%
   plot # the same as above
dend %>%
   branches_attr_by_labels(c("123", "126", "23", "29"), "any") %>%
   plot

dend %>%
      branches_attr_by_labels(c("123", "126", "23", "29"),
      "any", "col", c("blue", "red")) %>% plot
dend %>%
      branches_attr_by_labels(c("123", "126", "23", "29"),
      "any", "lwd", c(4,1)) %>% plot
dend %>%
      branches_attr_by_labels(c("123", "126", "23", "29"),
       "any", "lty", c(2,1)) %>% plot

# }