click_rotate {dendextend}	R Documentation

Interactively rotate a tree object

Description

Lets te user click a plot of dendrogram and rotates the tree based on the location of the click.

Code for mouse selection of (sub-)cluster to be rotated

Usage

click_rotate(x, ...)

## Default S3 method:
click_rotate(x, ...)

## S3 method for class 'dendrogram'
click_rotate(x, plot = TRUE, plot_after = plot,
  horiz = FALSE, continue = FALSE, ...)

Arguments

x
a tree object (either a dendrogram or hclust)

...
parameters passed to the plot

plot
(logical) should the dendrogram first be plotted.

plot_after
(logical) should the dendrogram be plotted after the rotation?

horiz
logical. Should the plot be normal or horizontal?

continue
logical. If TRUE, allows the user to keep clicking the plot until a click is made on the labels.

Value

A rotated tree object

Author(s)

Andrej-Nikolai Spiess, Tal Galili

See Also

rotate.dendrogram

Examples

# create the dend:
dend <- USArrests %>% dist %>% hclust("ave") %>%  
as.dendrogram %>% color_labels

## Not run: 
# play with the rotation once
dend <- click_rotate(dend)
dend <- click_rotate(dend, horiz = TRUE)
# keep playing with the rotation:
while(TRUE) dend <- click_rotate(dend)
# the same as
dend <- click_rotate(dend, continue = TRUE)

## End(Not run)

[Package dendextend version 1.9.0 Index]