dist.dendlist {dendextend}	R Documentation

Topological Distances Between Two dendrograms

Description

This function computes the Robinson-Foulds distance (also known as symmetric difference) between two dendrograms. This is the sum of edges in both trees with labels that exist in only one of the two trees (i.e.: the length of distinct_edges).

This function might implement other topological distances in the future.

Usage

dist.dendlist(dend, method = c("edgeset"), ...)

Arguments

dend
a dendlist

method
currently only 'edgeset' is implemented.

...
Ignored.

Value

A dist object with topological distances between all trees

See Also

distinct_edges, dist.topo, dist.multiPhylo, treedist,

Examples

x <- 1:5 %>% dist %>% hclust %>% as.dendrogram
y <- set(x, "labels", 5:1)

dist.dendlist(dendlist(x1 = x,x2 = x,y1 = y))
dend_diff(x,y)

[Package dendextend version 1.9.0 Index]