dist_long {dendextend}	R Documentation

Turns a dist object to a "long" table

Description

Turns a dist object from a "wide" to a "long" table

Usage

dist_long(d, ...)

Arguments

d
a distance object

...
not used

Value

A data.frame with two columns of rows and column names of the dist object and a third column (distance) with the distance between the two.

Examples

data(iris)
iris[2:6,-5] %>% dist %>% data.matrix
iris[2:6,-5] %>% dist %>% as.vector
iris[2:6,-5] %>% dist %>% dist_long
# This can later be used to making a network plot based on the distances.

[Package dendextend version 1.9.0 Index]