distinct_edges {dendextend}	R Documentation

Finds distinct edges in one tree compared to another

Description

Finds the edges present in the first tree but not in the second

Usage

distinct_edges(dend, dend2, ...)

Arguments

dend
a dendrogram to find unique edges in

dend2
a dendrogram to compare with

...
Ignored.

Value

A numeric vector of edge ids for the first tree (dend) that are not present in the second tree (dend2).

Source

A dendrogram implementation for distinct.edges from the distory package

See Also

distinct_edges, highlight_distinct_edges, dist.dendlist, tanglegram distinct.edges

Examples

x <- 1:5 %>% dist %>% hclust %>% as.dendrogram
y <- set(x, "labels", 5:1)
distinct_edges(x, y)
distinct_edges(y, x)
dend_diff(x, y)
# tanglegram(x, y)

[Package dendextend version 1.9.0 Index]