fac2num {dendextend}	R Documentation

Turns a factor into a number

Description

Turning a factor into a number is not trivial. Using as.numeric would only return to us the indicator numbers and NOT the factor levels turned into a number. fac2num simply turns a factor into a number, as we often need.

Usage

fac2num(x, force_integer = FALSE, keep_names = TRUE, ...)

Arguments

x
an object.

force_integer
logical (FALSE). Should the values returned be integers?

keep_names
logical (TRUE). Should the values returned keep the names of the original vector?

...
ignored.

Value

if x is an object - it returns logical - is the object of class dendrogram.

Examples

x <- factor(3:5)
as.numeric(x) # 1 2 3
fac2num(x) # 3 4 5

[Package dendextend version 1.9.0 Index]