which_leaf {dendextend}	R Documentation

Which node is a leaf?

Description

Gives a vector as the number of nodes (nnodes), which gives a TRUE when a node is a leaf.

Usage

which_leaf(dend, ...)

Arguments

dend
a dendrogram dend

...
ignored.

Value

A logical vector with the length of nnodes, which gives a TRUE when a node is a leaf.

See Also

noded_with_condition, is.leaf, nnodes

Examples

## Not run: 

library(dendextend)

# Getting the dend dend
set.seed(23235)
ss <- sample(1:150, 10 )
dend <- iris[ss,-5] %>% dist %>% hclust %>% as.dendrogram
dend %>% plot

which_leaf(dend)

## End(Not run)

[Package dendextend version 1.9.0 Index]